Bug fix:
In coda_venus_readdir(), the handling of error return of linux kernel
function filldir() is corrected: return of EINVAL simply means we have
use up all the user allocate space for getdents(3), it is not really
an error: the rest of the dirents can be read in next round, so the
return value of filldir() should not be propagated up for
coda_venus_readdir().

This fix the previous bug that on linux that 'ls -l' sometime failed
on certain directories with long content (e.g ls -l /coda/usr), but not
on other directories with short content (e.g. ls -l /coda).  It also
fixes other noted problem with getdents(3) (e.g. repair(1) said
"Couldn't stat current working directory").

-- clement