Bug fix:
There was a race possible in log_recover(): first_rec_num is not
protected by mutex lock of dev_lock, so before log_recover() writes
out first_rec_num together with prev_log_head et. al, it is possible
that someone else (e.g. a record flush) may do a write_log_status()
and written out partial data in the log status block on disk.  Should
the program fails at that point, meaning that truncation is not
completed and status_update() is not carried out, inconsistent state
will be left on the log status block on disk.  This was the bug which
crashed rossini on Apr 25, 1997.  (c.f. post on coda.bugs on May 6,
1997).

Solution: delay the update of in-memory copy of first_rec_num (of log
status) until status_update(), where under the protection of the mutex
lock dev_lock, in-memory copy of first_rec_num is updated, together
with zeroing of prev_log_head et. al and is written out to disk.

-- Clement