NetBSD kernel bug work around:
NetBSD has a kernel bug that will panic if we try to read from a raw
device and copy it to address on or above 0x10400000.  This is known
to be a problem with vm_fault() of NetBSD kernel that panics when it
finds that the pte (page directory table entry) does not exist in page
dir table (instead of trying to create it). Before that is fixed, we
work around it by manually touching one byte of address space of every
pte's that we'll need.  This will get the pte created and we'll be
fine.  This is proposed by rvb.

  -- clement