Bug fix:
V_GID should be defined as 65534 rather than (vgid_t)-2, since the
latter evaluates to 4294967294 (vgid_t is "unsigned int"), which is not
our intention.

Benefit:

This fixes the problem with chown with gid 65534 (the gid that venus
runs under).  For example, "cp -p" on linux got "Permission denied
message" even though the user was authenticated.  The permission error
was not with uid but with gid as "cp -p" on linux did a
chown(foo, UID, 65534) as the last step.