Bug fix:
On class olist_iterator:

Do we allow people delete a object while it is returned by class
olist_iterator ?  I believe so.  For example, in vrtab::clear()

    ohashtab_iterator next(*this, (void *)-1);
    ...
    while (vre = (vrent *)next()) {
        ...
	delete vre;
    }

(ohashtab_iterator make use of olist_iterator)

If it is true, then this should be accounted for when we design the
operator () for olist_iterator.  Namely, we should *save* the pointer to
the next olink in list before return the pointer to current olink, as
the caller may delete the current olink right the way.  In that case,
in the next iteration, we will refer to memory region that has already
been freed !

This is implemented in this fix.