Better debugging support:
Added an infinite sleeping loop in zombie() so that gdb attach to a
failing server process is now possible.  This is for non-mach platform
only, mach platform already had the task_suspend() in place.  Non-mach
platform previously let the signal handler zombie() to return, which
was not entirely correctly.

This should also fix the bug on NetBSD where sometimes
there were infinite messages of something like this

> 19:43:17 ****** FILE SERVER INTERRUPTED BY SIGNAL 11 CODE 6 ******
> 19:43:17 ****** Aborting outstanding transactions, stand by...
> 19:43:17 Uncommitted transactions: 0
> 19:43:17 Uncommitted transactions: 0

appeared on SrvLog and eventually filled up the whole disk.