Bug fix:
It was an incorrect use of RVM.  The routine rvm_create_segment()
specifically requires the second argument DevLength to be zero if the
segment device is a file.  This is stated explictly in RVM manual.
However, in the routine Recov_CreateSeg(), the following line with +
sign was missing, meaning that whatever garbage on the stack would be
used as the value for the 2nd argument dummy (length of the segment
device), hence all the troubles with RecovInit() ...

    rvm_offset_t dummy;
+   RVM_ZERO_OFFSET(dummy);	/* VenusDataDevice is file, must zero dummy */
    rvm_return_t ret = rvm_create_segment(VenusDataDevice, dummy,
					   &Recov_Options, nregions, regions);

This bug did not show up often.  This is probably because it was only
involved when veni were brainwiping of rvm store (with the -init
switch), which did not happen often.