Minor fix:
The ticket-granting-ticket, required by krcp, has a life time of only
5 minutes.  If there are a lot of hosts to contact, and/or some of
them are dead and take time to time-out, five minutes may not be
enough for the whole process of getting VolumeList from hosts.
Therefore we need to, from time to time, use "klist" to check if we
still have the ticket, and request one if we don't.