Merged Hiroshi Inamura <inamura@isl.ntt.co.jp>
Hi,

Here is a patch of the diff between my source and CMU coda tree on 14
Jul 97. This is the version I use now and successfully build the whole
Coda on. I am using Coda on FreeBSD 2.2.1-RELEASE.

To use the diff on FreeBSD, be sure to turn off 'cluster write' flag
in /usr/src/sys/ufs/ffs/ffs_vnops.c. when rebuild the kernel. To do
this, following patch can be used. If do not, a large file is not
written correctly in Coda name space.  Amazingly, this happens only
with venus, not with potemkin.  I am tracking this.