Bug fix:
I added a call to setrlimit() to give coda server the maximum limit on
data size.  We found that coda servers on NetBSD, from time to time,
could not get enough memory by malloc.  We believe it is due to the
fact that default resource limit on NetBSD is much more restricted
than that of Linux.  Hope this help.