Improvement:
I modify the zombifying routine FatalSignal() so that, even in linux,
it can print meaningfull backtrace.  I do this by using sigsuspend()
directly rather than using sleep() (which call sigsuspend()).  This
way, I can avoid an in-kernel __sigmask, which was found to be in
inaccessible memory region and was the obstracle for meaningful
backtrace to be printed.