Bug fix:
Now, we can attach to a zombified server and still get meaningful
backtrace.  This was not possible before on Linux.  I use similiar
technique posted earlier (which has been applied to Venus): Instead of
using sleep() to put the process to sleep, I use sigsuspend().

I also fix one bug in the signal handler zombie().  In Linux, the
signal handler are *not* passed the second and third arguments, or
code and scp respectively, as in NetBSD.  Trying to use them within the
signal handler may casue segfault, so we might have had segfault in a
segfault handler...