*** Important ***
This bug fix needs a new RVM log version, if you have any existing RVM
logs, you must truncate them and then switch to logs of new format.
See the rvm-src/rvm/README.log1.4 for the transition procedures.

*****************

Bug fix:

I fixed an once-in-a-million-time bug of rvm.  This bug happened as
follows.  Whenever there is a log wrapping, RVM system has to scan
backward from the log end to locate the log-wrap marker (which may
located not exactly at the end of log but certain bytes before the log
end).  The scanning was done by trying to locate the struct_id field
of wrap marker, which contains the constant numeric value 25
(log_wrap_id).  The bug surfaced when the timestamp.tv_usec field of
the wrap marker (which can take value from 0 to 999,999) was so
luckily to take the value 25 !  The timestamp field was scanned
*before* the struct_id so indeed it was possible to mistake the
timestamp.tv_usec field as the struct_id field.  Misinterpreting
log-wrap marker means misinterpreting the entire log, and RVM could do
nothing but die.  The solution that I propose is to put another
struct_id2 field at the *end* of wrap marker.  This way, RVM will
always examine struct_id2 before it examines other field of the wrap
marker, so this nasty bug could never happen again.