A lot of patches for proper initialization of buffers, and some general cleanup.
The problems encountered all are of the following form:
 a token is an array of 8 bytes.
 a password is a null terminated string, not longer than 8 bytes

RPC2 did something evil (among many other evil things in the code):

It says: a token is a string, with padding to reach a multiple of 4
bytes. So it sends tokens to the server with _uninitialized_ padding.
As a result the characters following a password shorter than 8 (in fact 7)
characters would get mauled.  The auth2 server now manages this.

I have not been able to track down the "hang" in clog.  It happens in
the RPC2_Newbinding call on the client and the GetRequest call on the
server.