Returning an RPC2_NAKED instead of RPC2_FAIL when a token expired.
RPC2_FAIL was essentially processed as a timeout, i.e. servers marked
down, failed operation got logged. But this has the annoying side
effect that the user is totally unaware of the tokenloss and Venus
silently switches from fully connected to disconnected operation.

By returning RPC2_NAKED, the server is not marked down, but the
connection to the server is rebound (1st fails on the expired token,
2nd as System:AnyUser) and the operation is retried. If System:AnyUser
is not allowed to perform the operation EACCESS is returned to the
user. Who then hopefully will obtain a new token and redoes the
operation.