Previous fix avoided the crash, but was bogus, as we didn't collapse the expanded namecontexts anymore. I still haven't figured out why we get a destroyed namecontext from the prioq, but discovered that we can skip it whenever we hit one, and that is does get cleaned up correctly. Currently the code logs the following message,     Hoard daemon got dead or dying namecontext 0x818ad18 , skipping it!
After `fixing' the namectxt::~namectxt crash, I discovered that the
hoardlist was never re-expanded when the user regained a token.
Demotion wasn't strong enough due to some optimizations in MetaExpand.
By clearing out the expander_fid/expander_VV of a namecontext during
demotion, we always force a re-expansion after a demotion. It slows
down the first hoardwalk following a demotion, but namecontexts are
only demoted by the user losing/regaining a token and after an error
during a hoardwalk/namecontext-expansion.