Removed all code where we attempted to lock a directory, as it is impossible to get an exclusive lock on a directory using the posix fcntl locking scheme. The simple reason is that the fd has to be opened O_WRITE, but it is not possible to to obtain an open fd for a directory when the O_WRITE flag is specified.
While removing the directory locks I found some strange lock uses in
other places as well. The volutil.lock probably doesn't lock at all:

    fd = open("volutil.lock");
    lock(fd);
    close(fd);
    ....
    fd = open("volutil.lock");
    unlock(fd);
    close(fd);

As far as I can tell, the lock was already dropped after the first
close. Furthermore, the locking of /vicep* partitions is messed up, it
allows multiple calls to lock the partition, done once for every
volume?, but clears the lock the first time unlock is called. This
code is now not active as it attempted to lock the /vicep* mountpoint.