Modified venus to use a unix domain socket for the mariner port, the tcp port defaults to off, but can be turned on by giving venus the -MarinerTcp flag.
When unix domain sockets are not available venus will default to using
the tcp socket, but it will also print a warning. The port can be
deactivated by starting venus with the -noMarinerTcp flag.

Codacon and spy have been modified to connect to the unix domain
socket by default. However, to connect to the tcp port pass them -tcp.