Although replacing strbase with a version that fixes the gcc-3.2 compile warnings and 'should' compile to correct code, it in fact doesn't.
It works fine when we're dereferencing from a containing class or struct
to the container class, but breaks when the container is a struct, which
happens to be the case for the VnodeDiskObject on the server. Reverting
strbase, I rather live with warnings than servers that don't start
correctly.

Also after the last patch I could removed any trace of the VSGDB in the
Coda server. Which means that now only createvol_rep is using the VSGDB
and thus we can start looking at dropping the VSG identifiers completely
(neither client nor server really cares about them anymore).