Eureka, I think I found the FreeBSD mountpoint traversal bug.
The Linux kernel uses a directory cache, so we never see lookups for
'.' and '..' At some point I moved the '..' lookup from fsobj::Lookup
to vproc::lookup which simplified the code.

However, it looks like the code was a bit too eager following
mountpoints and would do the same thing for the VenusRoot object i.e.
whatever is mounted at /coda. As a result venus will probably either
die, or the FidToNodeId won't recognize that we just hit the top of
our tree and substitute the real inode number of the /coda mountpoint.

If this is right, FreeBSD (and NetBSD) should be ready to roll.