Did some productive bug-hunting,
- Unpacking for BoundedBS data was too strict and would drop incoming
packets, this was causing failed runt resolution. Only affects servers
build against CVS versions of RPC2. This problem was in the rp2gen
generated stub code so it will require a rebuild of client and server.

- Avoid crashing when trying to send a multicast rpc to an empty mgrp.
This can happen during resolution when hosts become unreachable, or
drop packets due to something like the above bug. This is an old bug
that could only hit in cases where a sequence of operations is
performed on the same mgrp, typically resolution. Bug is in the
library, so clients/servers can be fixed by just dropping in the new
library.

- Silly change, dmalloc doesn't like malloc(0), there aren't many
places where we call this anyways, except for where we send a multirpc
without arguments. Quick fix was to just add 1 to the requested size.

- Horrible bug, that has been in the multirpc string packing function
since the beginning of time.

    int **p; *(char *)((*p)+strlen(string)) = '\0';"

It took me a while to actually find this because the result was very
unpredictable memory bytes being zapped to zero when a client was f.i.
reintegrating. This is in the dynamic library and so existing clients
and server can be fixed by simply dropping in a new shared library.

- Potential bug, we weren't padding after packing an RPC2_Secretkey.
Not really a problem because the keylength happens a multiple of the
padding size, but if anyone ever decides to change the keysize to
something odd.