- When regular directory resolution tells us that all object are already marked as in conflict, mark them again. That way at least the version vectors will get synced up and repair can actually commit a fix for the problem without it failing in the subsequent resolution attempt on the versionvector differences.
- During runt resolution, when fetching the object from a non-runt
server, ignore the inconsistency flag, normal resolution will kick
in after the runts are updated and will fail when they are not
updated. Forcing us to resolve the parent which instantiates the
object on which we then do often unsuccessful log-based resolution.

- When forcing updates to runts, don't send a copy of the data to
replicas that already have a non-runt copy.

- coda-client-logrotate was looking for the wrong 'pid_file' entry in
venus.conf so logs didn't get rotated.

- I didn't see a reason why we shouldn't allow demotion for volume
mountpoints. This is especially useful when repairing conflicts at
or in the root of a volume.

- Update an object's versionvector when we receive new 'status'
information even when it is not in a replicated volume. Some tools
like repair actually use the VV from objects in non-replicated volumes
and the server rejects the repair operation when it doesn't match.
This explains why an 'impossible' repair often is simple from another
client, the other client simply didn't have a bad VV cached.

- Don't bother revalidating dirty objects with ViceValidateAttrs.

- Don't look up names that start with '@' in the fake rootvolume, this
avoids useless DNS lookups when some realm's root has a conflict.

- When reintegration fails, only dump the part of the CML that we
just tried to reintegrate in the log instead of all (thousands) of
entries.

- Clear the partial reintegration handle when we hit a conflict, this
way we avoid getting stuck in an endless loop trying to close the
reintegration when we attempt to do partial reintegration on top of
a server-server conflict.