Red and yellow zones for writeback cached operations.
  redzone = free CMLs < MaxWorkers or free(able) Blocks < 6%
  yellowzone = free CMLs < 12% or free(able) Blocks < 25% (&& !redzone)

If we are in the yellow zone we just slow down write operations by
pausing approximately one second. If we are in the red zone we start
blocking mutating operations in write-disconnected (and disconnected)
volumes.

Maybe we should only block if we are 'redzoning' when write-
disconnected. We might also want to kick off a forced reintegration
when we get into the yellow zone, but this seems to work ok for now.