Fix venus crash bug reported by A. Wegele.
When a file was opened O_CREAT | O_RDONLY, the open implicitly turned
the O_RDONLY into O_RDWR, but close never sees the O_CREAT part and we
ended up with a Writers refcount on the object that was never dropped.
Then when the object is unlinked from the same client, an assertion
triggered because the object had no 'openers' but one remaining
'writer'.

Also moved DiscardData in fsobj::Release up a couple of lines to avoid
a NULL ptr dereference on the volume pointer when DiscardData fails.