Commit of ucontext.h patch.
- If ucontext.h is available we use makecontext to create threads and
swapcontext/setcontext to switch between those threads. If ucontext.h is
not available we try to use sigaltstack for thread creation and
sigsetjmp and siglongjmp for the thread switches. Finally if sigaltstack
is unavailable we fall back on the the old savecontext/returnto assembly
routines in process.S to initialize threads. The result even seems to
work on cygwin.
- Also includes some cleanups for the pthread variant of LWP, now
liblwp_pt works on cygwin as well.

This library is binary (drop-in) compatible with previous releases.
However I cleaned up the headers a bit to solve internal confusion
(hidden by casts) about the return types of callback functions.
Applications that try to build against this will see some compiler
warnings and errors that require some trivial fixes, I will commit the
necessary fixes for RPC2, RVM and Coda later on.