From: Greg Troxel <gdt.ir.bbn.com>
lwp head fails to build on NetBSD/i386 1.6.2 because it contains
declarations after code.  Patch appended to fix it.  With this,
lwptest-static runs fine.  But, this is w/o ucontext but w/
sigaltstack.

But, the same static binary copied to a 2.99.15 system and a 2.0
system segfaults, apparently in errno() in libc.

(gcc 3 really needs an option to disallow C99isms; lots of software is
declaring variables after statements these days!)