Attempt to fix conflicts with the official Debian package.
The librpc24 package in the official Debian archive has the name
librpc2-4. However the dev and tools packages have identical names to
ours. Because these packages have an additional '-debianX' component for
their version, apt believes they are more recent and will try to
upgrade. However they pull in librpc2-4 which conflicts with our
librpc24, but that package doesn't declare a conflict, as a result the
install fails. The only way to break out of the conflict is to remove
librpc24 and the Coda client/server packages.