Hooks in NewBinding to allow for an alternate connection setup.
Set an RCP2SEC_CAPABLE flag on the INIT1 packet, existing servers simply
ignore this flag and continue with the existing handshake. However this
flag allows a security enhanced rpc2 server to complete an alternate
handshake.

Also added a socket listener return code so that we can skip the rest of
RPC2_NewBinding if the security handshake completed while the client was
waiting for the INIT2 response.