Added secure_sendto and secure_recvfrom.
Wrappers around sendto and recvfrom that optionally encrypt or decrypt
based on a 'security association'. The implementation follows the
recommendations for incoming and outgoing packet processing from
several IPsec related RFCs, specifically
    RFC 2406 - IP Encapsulating Security Payload (ESP)
    RFC 3948 - UDP Encapsulation of IPsec ESP Packets

This code currently does not do all that much since it does not include
any functionality to set up and initialize the security associations.