Bump the maximum packet size up to 1452.
The problem is ValidateAttrsPlusSHA. The compiled default for Coda
clients is to validate 21 fids at a time, which results in a 1376
byte packet. Since I want the new RPC2 library to be a drop-in
compatible replacement we pretty much have to support that packet
size.

In the next Coda release I'll probably reduce the default setting
to 15 fids, which makes the packet small enough for the IPv6 minimum
MTU. At some point the ValidateAttr call should be cleaned up, drop
the 'getattr' bits and use a smarter FID/VV packing to eliminate the
redundant data (volumeid, null VV entries, VV.flags).