Avoid RPC2 handles < 256 and move HandleInit1 out of the main path
Since we use RPC2 handles as security identifiers, we avoid the
IANA reserved range which is handled differently by secure_recvfrom.

Also moved HandleInit1 out of the common DecodePacket path, the init1
packets are the only ones which do not have an established connection
this simplifies the logic in DecodePacket.

Finally, avoid wasting time trying to shrink packets that would be
dropped later on.