Use a real password-based key derivation function
My ad-hoc attempt was broken because it was too fast. In the real
world people use a similar PRNG to derive a key, but add a lot of
unparallelizable iterators to make brute-forcing more expensive.
So here is a PBKDF implementation based on PKCS #5 version 2 which
is internally using AES-XCBC-PRF-128 as the pseudo random function.