Merged 64-bit fixes from Brett Lymn
Change VnodeDiskObject.inodeNumber to a union of a pointer to the
RVM representation for directories and an container file identifying
integer for files and symlinks. Initializing realm::realmid with a
global counter value, the 32-bit truncated pointer is not a useful
uniquifier on a 64-bit system. And various other cleanups.