Fix RPC2 binding on amd64.
The Init1 packet contains a FakeBody which in turn has a char *. We
don't actually send a pointer to the other side, but it is used as an
anchor to copy a char array. The problem is that the pointer is larger
on 64-bit which breaks the layout of the Init1 packet and rpc2 binding.
Fixed it by changing the Init1 structure to not contain the pointer.