SmallDiskVnodeObject is larger on a 64-bit arch.
The size is hardcoded (i.e. not sizeof(SmallDiskVnode)) to catch
potential alignment issues as a result of compiler/ABI changes. However
the test to see if the hardcoded number and the sizeof still match
probably got lost somewhere. So we ended up clobbering the diskvnode
structure on a 64-bit arch whenever we changed an ACL.