Add fail_delay_tx/fail_delay_rx support.
delay, color = fail_delay_xx(addr, lengthofpacket, color)
Where addr is where the packet is sent to/received from, length is the
size of the packed (not including ip/udp header overhead) and color is
the current packet color. The returned color value is used to tag the
packet with a new color value. If delay is nil the packet is dropped.
when delay is > 0 the packet transmission or reception is delayed by
the specified number of microseconds.

Load standard libraries in the Lua environment and initialize
RPC2_RETRIES and RPC2_TIMEOUT variables.