Fix the AES-CCM initializer.
We were not masking off the correct (flag) byte when initializing the
counter block. Actually discovered the problem when a big-endian PPC
would not complete a handshake with a little-endian x86. Initialization
on the powerpc actually ended up being the correct one.

Also took out the union to avoid any chance of confusion in the future.