Fix compatibility after AES-CCM initialization change.
The incorrectly initialized counter block in the v0 implementation of
AES-CCM is most likely weaker since it cleared 5 bits of the IV. So we
should try to establish the best possible security when we encounter a
combination of old and new clients and servers.

v0 client, v1 server - The server will recognise that the client is a v0
implementation and tell the client to use AES-CBC encryption for data it
sends to the server. The v0 client assumes that the server chose the
best possible algorithm and will also request AES-CBC encryption for any
further server-to-client traffic. (we simply avoid AES-CCM)

v1 client, v0 server - Server will request AES-CCM encryption and report
it's version number as 0. The client has no choice for client-to-server
traffic, but it does have a choice for server-to-client traffic, so it
tells the server to use AES-CBC for any data that is sent to the client.