Remove libfail/filcon
We can control network failure and bandwidth shaping from dynamically
loaded Lua scripts so we don't need libfail and filcon. We specifically
didn't link the Coda clients and servers with libfail because the fail
filters could be set by sending packet to the rpc2 port, which is not a
good idea for publically available servers.