Lower number of piggybacked attributes in ValidateAttrs to 15.
This brings the packet to below the IPv6 MTU, which is good. However
we need to send about 33% more RPC operations to validate a cache.
Although we have validatevols to efficiently revalidate a large client
cache when there are not many updates on the server, we still need a
better/more efficient solution to quickly revalidate large, and
relatively active, volumes.