Increase directory buffer size in resolution coordinator.
When we are fetching and comparing the directory data from all replicas
the coordinator allocated enough space for the directory and ACL data.
However the resolution clients include the quota values when we try to
resolve a volume root.

As a result, when the volume root directory is 256KB, we ended up being
8 bytes short in the receive buffer. As a result sftp would fail, the
connections were killed off and we failed to mark the directory as
successfully resolved.