Use the non-blocking /dev/urandom device.
We grab 32-48 bytes of random data to initialize our random number
generator. This isn't really a problem for long running processes
like codasrv and venus, but starting a lot of short lived ones such
as clog and volutil quickly drains the available entropy in a system.
The resulting program startup delay results is often misdiagnosed as
a network/dns/rpc2 problems and can lead to denial of service when
using pam_coda for logins or failing backups.