Add allow_backfetch option
Backfetching during reintegration may save unnecessary data transfers
when reintegrations are interrupted, but...

- Interrupted reintegrations are (or should be) rare.
- All servers will independently perform the backfetch, so we have to
  send the file N times, and can't even migrate to a more efficient
  multicast/broadcast scheme.
- Volume is locked on both client and server during the reintegration,
  the additional data transfer makes the reintegration RPC take longer.
- Backfetches use the callback connection, so updates for unrelated
  volumes can get blocked if we need to send a callback break to the
  reintegrating client.
- The initial handshake for callback connections is based on a known
  key and as such they are not as secure as normal client->server
  connections.

All in all it seems better to not use backfetches at all and rely on
partial (trickle) reintegration to push a store to a server and then
use resolution to propagate the update to all other replicas. This
option can be used to re-enable the old behavior.