Only set d_inodemod on directories
If we're renaming a file we use AddChild which was unconditionally
setting d_inodemod. This happened to overlap with f_sinode on 64-bit
so rename incorrectly dropped container file 1 whenever a file was
renamed. It also affected CML_Remove, however as we unlink the file
we overwrite the bad f_sinode value.