Fix retryinterval calculation.
When estimating the time to wait for the next packet, we first establish
a base to make sure we do not exceed the user specified maxrtt value.
The C implementation diverged from the design of the rpc2-rtt-adaptive
lua script and we ended up waiting twice as long. This fix brings the C
implementation back in line with the original specification.