Remove (most) of the __BSD44__ defines
Really we are all POSIX now, we don't really need many OS specific
defines. The only major differences are when we have to mount /coda
and how to get system wide statistic counters on the server (for
cmon/smon).

If there really is a missing or required header we should use an
autoconf test.