Fix errno translation.
We didn't actually use RPC2_S2RError, and only used R2SError when
collecting MultiRPC reponses. So when the errno's were changes we
actually ended up breaking the on-the-wire representation.

The only advantage was that most on-the-wire errors are identical
with the system errno values (at least on an i386 based Linux kernel).
Clearly we started seeing failures once we started using a mix of
other systems on our production servers.