In some cases SFTP is sending the same data multiple times.
One side of the SFTP transfer is passive and will always respond
to incoming packets but will not automatically retransmit. The other
(active) side also sends reponses based on incoming packets, but
also triggers retransmissions after a timeout. The problem is that
if we have no packet loss, but only delay, any timeout-retransmission
will effectively start a concurrent transfer of the same data.

So the active side should in some cases choose not to send more data
(or a new ack) when we receive an ACK or ACKME flagged packet. The
changes in this patch try to introduce some of this throttling. The
side that is sending data (receiving ACKs) will avoid resending
packets as long as we are receiving timely responses from our peer.