SFTP tweaks
Don't include 100msec delayed ack time in SFTP roundtrip estimates.
Reduce duplicate suppression to 4. The default was 16 which with the
default 10 retransmission retries meant we never recovered from a lost
data or ack packet on the client->server path.
Be more aggressive with resending worried packets.
However, only resend data if we hit a timeout.