Fix reaping of dead mariner threads.
If a thread is deleted while it has a queued timer entry, the timer is
not removed. As a result we will wake up the dead thread and it crashes
the client as it tries to requeue it self. Committing suicide ends up
being considerably safer because we know we don't hold any locks or have
any events we are waiting for.