Move per volume store identifier/uniquefier to a global structure.
The server uses the store uniquefier as a way to discover replayed
reintegration attempts. However we stored it as part of the per-volume
data structures which are destroyed when there are no more objects for
that volume in the cache. As a result any futher reintegrations after
this would be dropped on the floor.

This patch moves the uniquifier value to a single global which is
persistent across venus restarts and independent from individual volume
teardown and creation events. Fixes bug #1638.