Allow a client to disable replay detection.
If the store uniquifier value is less than INT_MAX the server performs
replay detection which avoids certain types of conflicts. If the value
exceeds INT_MAX detection is disabled. This way a client can indicate
to the server that it does not want replay detection, but it can still
work correctly with an old server which expects strictly increasing
store identifiers.