Packaging updates
If we don't build or install something from a subdirectory the
directory probably does not need a Makefile.
Add uninstalled pkg-config file which might become useful if we want
to build everything in place. Moved pkg-config files to the top-level
directory to make them easier to find for other packages.