Follow lwp's API change.
LWP_MwaitProcess prototypes became const void **, but the compiler
won't cast a void** because it can in some way result in code being
able to change a const void object without a compiler warning. So
the LWP api change broke the venus build, but this change breaks
the build against older versions of LWP. Interestingly the binary
interface remains unchanged so the library version is still the same.