Combined/cleaned up several mrpc_SendReliably code-paths
If the rpc of the last connection entry timed out and it had no recent
sftp traffic we would clear the hopeleft flag and abort too early.

The global time-bomb was assigned to the last slot of the pending array,
if any of the connections had a setup problem, LWP_MWaitProcess would
not notice if this timer expired.