Precalculate retry timeout values.
Rename per-connection keepalive to 'timebomb', because that is what
it really has been used for (kill long running rpc2 operations, f.i.
during binding).
Precalculate preferred retry intervals, but only use them when the
estimated roundtrip time is shorter. So retries really become a best
case scenario and we automatically wait longer if roundtrips are long.
Of course this means that we could wait considerably longer, hence the
timebomb value.
We're not actually using the timeout on regular RPC2 calls yet, I'd
first like to see the effect of the new retry interval calculation
and possibly lower the default timeout value first.