gcodacon: Add command-line option to disable notifications
Coda status changes are normal and happen frequently, so we really shouldn't
be showing an alert every time they do.  The user can disable these alerts,
but must do so manually on every gcodacon launch.  Provide a command-line
option to disable popups so that the user doesn't have to do it through the
GUI.  (It would be better to disable popups by default, but that would
surprise existing users.)