Revert "Grab shared locks for read operations."
When we are unable to grab a shared lock, the operation returns
an error to the client which retries the operation. However in
various places, such as server resolution, the retry failure is
unexpected and handled as fatal, which leaves the object marked
as an unresolvable conflict.

It also introduced non-fairness and potential starvation issues
when many clients retried their requests.

This reverts commit 8c4fdbb727c7476c4bce298d4205667f65685dc4.