Check for sufficient space to log a transaction.
When checking if enough space is available to write a transaction
to the log we failed to account for the size of a log_wrap_t record,
but we did reserve room for this that when writing each individual
range. As a result we could run into the log head. In this case it
is assumed we hit end of file and a log_wrap is initiated which would
trigger an assertion when we're not actually near the end of the log.